Extract death functions into module
